Initial mavenization
